refactor: 상단 메뉴탭 템플릿 view 폴더로 이동함
