+++ src/js/view/initScreenTemplate.js
